Don't rewrite unchanged NFO files
When computing the NFO contents, we will rewrite the file even if everything is unchanged. This causes the file updated time to change and increases load on the filesystem (and underlying drives)

Add `ComputeHash` to compute the _SHA256_ hash of the file-on-disk and the stream we've just created and if the same just returns.

Also now passes the `cancellationToken` down to the stream copy so we abort faster if cancelled.

Fixes #12978